migrate file-copying functions from util to runner/copier
